Created PlotPeakByLogValueHelper
