+++ Framework/CurveFitting/inc/MantidCurveFitting/Algorithms/PlotPeakByLogValue.h
@@ -42,26 +42,26 @@ Required Properties:
 class DLLExport PlotPeakByLogValue : public API::Algorithm {
   /** Structure to identify data for fitting
    */
-  struct InputData {
-    /// Constructor
-    InputData(const std::string &nam, int ix, int s, int p, double st = 0,
-              double en = 0)
-        : name(nam), i(ix), spec(s), period(p), start(st), end(en) {}
-    /// Copy constructor
-    InputData(const InputData &data)
-        : name(data.name), i(data.i), spec(data.spec), period(data.period),
-          start(data.start), end(data.end), ws(data.ws) {
-      indx.assign(data.indx.begin(), data.indx.end());
-    }
-    std::string name; ///< Name of a workspace or file
-    int i;            ///< Workspace index of the spectra to fit
-    int spec;         ///< Spectrum number to fit
-    int period;       ///< Period, needed if a file contains several periods
-    double start;     ///< starting axis value
-    double end;       ///< ending axis value
-    API::MatrixWorkspace_sptr ws; ///< shared pointer to the workspace
-    std::vector<int> indx; ///< a list of ws indices to fit if i and spec < 0
-  };
+  // struct InputData {
+  //   /// Constructor
+  //   InputData(const std::string &nam, int ix, int s, int p, double st = 0,
+  //             double en = 0)
+  //       : name(nam), i(ix), spec(s), period(p), start(st), end(en) {}
+  //   /// Copy constructor
+  //   InputData(const InputData &data)
+  //       : name(data.name), i(data.i), spec(data.spec), period(data.period),
+  //         start(data.start), end(data.end), ws(data.ws) {
+  //     indx.assign(data.indx.begin(), data.indx.end());
+  //   }
+  //   std::string name; ///< Name of a workspace or file
+  //   int i;            ///< Workspace index of the spectra to fit
+  //   int spec;         ///< Spectrum number to fit
+  //   int period;       ///< Period, needed if a file contains several periods
+  //   double start;     ///< starting axis value
+  //   double end;       ///< ending axis value
+  //   API::MatrixWorkspace_sptr ws; ///< shared pointer to the workspace
+  //   std::vector<int> indx; ///< a list of ws indices to fit if i and spec < 0
+  // };
 
 public:
   /// Algorithm's name for identification overriding a virtual method
@@ -83,13 +83,13 @@ private:
   void exec() override;
 
   /// Get a workspace
-  InputData getWorkspace(const InputData &data);
+  // InputData getWorkspace(const InputData &data);
 
   /// Set any WorkspaceIndex attributes in the fitting function
   void setWorkspaceIndexAttribute(API::IFunction_sptr fun, int wsIndex) const;
 
   /// Create a list of input workspace names
-  std::vector<InputData> makeNames() const;
+  // std::vector<InputData> makeNames() const;
 
   /// Create a minimizer string based on template string provided
   std::string getMinimizerString(const std::string &wsName,
